page fault seems to work, but value is not written/read correctly

+++ headers/kernel/utils.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-void int_to_ascii(uint64_t, char*);
+void int_to_ascii(uint32_t, char*);
 void append(char s[], char n);
 void backspace(char s[]);
 int strcmp(char s1[], char s2[]);
